server files + dependencies
This is a prepack that contains the server files for both RPG and all of its dependencies.
This DOES NOT contain Metamod or Sourcemod.

Perform a fresh install of MM & SM, then install this prepack.
Edit your /configs/databases.cfg and replace the appropriate lines with your MySQL information.

Success!